crawler/requirements.txt
scrapy==2.11.2
scrapy-redis==0.7.3
fasttext-wheel==0.9.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.18.0
redis==5.0.1
